Created basic Web App and UI
Created a basic application that displays results obtained from the
classifier in txt format. The parser.py parses those results and send
the json objects to the server which then adds the data to a basic
SQLite database. The data from the database is then queried and used to
render the HTML which displays the results in a table.